水位可以填小数点— 运力调度：把预扣款去掉；单独搞个计划加油模块（含预扣-临调、计划加油-自有挂靠）；生成调度结果里面加个船长电话（承包人电话）— 船舶信息管理：最后3个日期不要必填，同时清空；干货船改成普通货船；营运类型选择集装箱后调度时不展—- 船舶信息管理支持可以根据不适航查询：不适航状态、原因（50字以上）—

+++ business/src/main/java/com/xingyi/logistic/business/bean/BaseWaterLevel.java
@@ -19,7 +19,7 @@ public class BaseWaterLevel extends BaseModelAndDO{
      * 水位
      */
     @NotNullEmpty
-    private Integer level;
+    private Float level;
 
     /**
      * 状态  1：启用  2：禁用
@@ -46,11 +46,11 @@ public class BaseWaterLevel extends BaseModelAndDO{
         this.name = name;
     }
 
-    public Integer getLevel() {
+    public Float getLevel() {
         return level;
     }
 
-    public void setLevel(Integer level) {
+    public void setLevel(Float level) {
         this.level = level;
     }
